requirementsNotFinal.txt
django
htmx
django-tables2
django-filter
django-htmx
crispy-bootstrap5
django-widget-tweaks